+++ src/main/java/com/telepathicgrunt/repurposedstructures/RSAddFeatures.java
@@ -329,14 +329,12 @@ public class RSAddFeatures {
         return false;
     }
 
-
     ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-    // MISC FEATURES //
+    // SWAMP TREE FEATURES //
 
     private static final ConfiguredFeature<?, ?> VANILLA_SWAMP_TREE = Feature.TREE.configure(DefaultBiomeFeatures.SWAMP_TREE_CONFIG).createDecoratedFeature(Decorator.COUNT_EXTRA_HEIGHTMAP.configure(new CountExtraChanceDecoratorConfig(2, 0.1F, 1)));
-    private static final ConfiguredFeature<?, ?> VANILLA_BOULDER = Feature.FOREST_ROCK.configure(new ForestRockFeatureConfig(Blocks.MOSSY_COBBLESTONE.getDefaultState(), 0)).createDecoratedFeature(Decorator.FOREST_ROCK.configure(new CountDecoratorConfig(3)));
 
-    public static void addMiscFeatures(Biome biome, String biomeNamespace, String biomePath) {
+    public static void addSwampTreeFeatures(Biome biome, String biomeNamespace, String biomePath) {
 
         // only exists in vanilla biomes
         if (RepurposedStructures.RSAllConfig.RSMainConfig.misc.hornedSwampTree && !biomeNamespace.equals("ultra_amplified_dimension") && biome == Biomes.SWAMP) {
@@ -346,7 +344,7 @@ public class RSAddFeatures {
         // can exist in modded biomes too
         else if (RepurposedStructures.RSAllConfig.RSMainConfig.misc.hornedSwampTree &&
                 (biome == Biomes.SWAMP_HILLS ||
-                        (RepurposedStructures.RSAllConfig.RSMainConfig.misc.addMiscToModdedBiomes &&
+                        (RepurposedStructures.RSAllConfig.RSMainConfig.misc.addSwampTreeToModdedBiomes &&
                                 biome.getCategory() == Category.SWAMP &&
                                 !biomeNamespace.equals("ultra_amplified_dimension") &&
                                 !biomeNamespace.equals("minecraft")))) {
@@ -355,10 +353,18 @@ public class RSAddFeatures {
             biome.getFeaturesForStep(GenerationStep.Feature.VEGETAL_DECORATION).removeIf(configuredFeature -> configuredFeature.config instanceof DecoratedFeatureConfig && serializeAndCompareFeature(configuredFeature, VANILLA_SWAMP_TREE));
             biome.addFeature(GenerationStep.Feature.VEGETAL_DECORATION, RSFeatures.HORNED_SWAMP_TREE.configure(DefaultBiomeFeatures.SWAMP_TREE_CONFIG).createDecoratedFeature(Decorator.COUNT_EXTRA_HEIGHTMAP.configure(new CountExtraChanceDecoratorConfig(2, 0.8F, 1))));
         }
+    }
+
+    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+    // MISC FEATURES //
+
+    private static final ConfiguredFeature<?, ?> VANILLA_BOULDER = Feature.FOREST_ROCK.configure(new ForestRockFeatureConfig(Blocks.MOSSY_COBBLESTONE.getDefaultState(), 0)).createDecoratedFeature(Decorator.FOREST_ROCK.configure(new CountDecoratorConfig(3)));
+
+    public static void addBoulderFeatures(Biome biome, String biomeNamespace, String biomePath) {
 
         if (RepurposedStructures.RSAllConfig.RSMainConfig.misc.boulderGiant && !biomeNamespace.equals("ultra_amplified_dimension") &&
                 ((biome == Biomes.GIANT_SPRUCE_TAIGA_HILLS || biome == Biomes.GIANT_TREE_TAIGA_HILLS) ||
-                        (RepurposedStructures.RSAllConfig.RSMainConfig.misc.addMiscToModdedBiomes && !biomeNamespace.equals("minecraft") &&
+                        (RepurposedStructures.RSAllConfig.RSMainConfig.misc.addBoulderToModdedBiomes && !biomeNamespace.equals("minecraft") &&
                                 ((biomePath.contains("giant") && biomePath.contains("taiga")) || biomePath.contains("redwood"))))) {
 
             // replace the boulders with our own
@@ -367,7 +373,7 @@ public class RSAddFeatures {
         }
         else if (RepurposedStructures.RSAllConfig.RSMainConfig.misc.boulderTiny && !biomeNamespace.equals("ultra_amplified_dimension") &&
                 ((biome == Biomes.SNOWY_TAIGA_MOUNTAINS || biome == Biomes.TAIGA_MOUNTAINS) ||
-                        (RepurposedStructures.RSAllConfig.RSMainConfig.misc.addMiscToModdedBiomes && !biomeNamespace.equals("minecraft") &&
+                        (RepurposedStructures.RSAllConfig.RSMainConfig.misc.addBoulderToModdedBiomes && !biomeNamespace.equals("minecraft") &&
                                 biomePath.contains("taiga") && (biomePath.contains("mountain") || biomePath.contains("hill"))))) {
 
             biome.addFeature(GenerationStep.Feature.LOCAL_MODIFICATIONS, RSFeatures.BOULDER_TINY.configure(FeatureConfig.DEFAULT).createDecoratedFeature(Decorator.COUNT_TOP_SOLID.configure(new CountDecoratorConfig(2))));
@@ -454,11 +460,10 @@ public class RSAddFeatures {
         }
     }
 
-
     ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     // TEMPLES //
 
-    public static void addTemplesAndPyramids(Biome biome, String biomeNamespace, String biomePath) {
+    public static void addTemples(Biome biome, String biomeNamespace, String biomePath) {
 
         if (RepurposedStructures.RSAllConfig.RSTemplesConfig.temples.netherBasaltTempleSpawnrate != 1001 &&
                 biome.getCategory() == Category.NETHER && (biomePath.contains("basalt") || biomePath.contains("blackstone")) &&
@@ -485,7 +490,13 @@ public class RSAddFeatures {
                 (biomeNamespace.equals("minecraft") || RepurposedStructures.RSAllConfig.RSTemplesConfig.temples.addNetherWastelandTempleToModdedBiomes)) {
             biome.addStructureFeature(RSFeatures.NETHER_WASTELAND_TEMPLE.configure(FeatureConfig.DEFAULT));
         }
+    }
 
+
+    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+    // Pyramids //
+
+    public static void addPyramids(Biome biome, String biomeNamespace, String biomePath) {
 
         if (RepurposedStructures.RSAllConfig.RSTemplesConfig.pyramids.netherPyramidSpawnrate != 1001 && biome.getCategory() == Category.NETHER &&
                 (biomeNamespace.equals("minecraft") || RepurposedStructures.RSAllConfig.RSTemplesConfig.pyramids.addNetherPyramidToModdedBiomes)) {
@@ -594,6 +605,7 @@ public class RSAddFeatures {
 
     /**
      * Will serialize (if possible) both features and check if they are the same feature. If cannot serialize, compare the feature itself to see if it is the same
+     * Doesn't actually serialize because I don't know how to do that in 1.16.1 lmao
      */
     private static boolean serializeAndCompareFeature(ConfiguredFeature<?, ?> feature1, ConfiguredFeature<?, ?> feature2) {