+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSMineshaftsConfig.java
@@ -18,6 +18,13 @@ public class RSMineshaftsConfig implements ConfigData {
     @Comment("Add the custom Mineshafts to modded biomes of the same categories/type.")
     public boolean addMineshaftsToModdedBiomes = false;
 
+    @ConfigEntry.Gui.Tooltip(count = 3)
+    @Comment("Add the ID/resource location of the biome you don't want"
+            +"\nRS's mineshafts to spawn in. Separate each ID with a comma ,"
+            +"\n"
+            +"\nExample: \"minecraft:ice_spikes,awesome_mod:awesome_biome\"")
+    public String blacklistedMineshaftBiomes = "";
+
     @ConfigEntry.Gui.CollapsibleObject
     public Spawnrate spawnrate = new Spawnrate();