+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSOutpostsConfig.java
@@ -14,6 +14,14 @@ public class RSOutpostsConfig implements ConfigData {
 
     public static class Outposts {
 
+        @ConfigEntry.Gui.Tooltip(count = 3)
+        @Comment("Add the ID/resource location of the biome you don't want"
+                +"\nRS's outposts to spawn in. Separate each ID with a comma ,"
+                +"\n"
+                +"\nExample: \"minecraft:ice_spikes,awesome_mod:awesome_biome\"")
+        public String blacklistedOutpostBiomes = "";
+
+
         @ConfigEntry.Gui.Tooltip(count = 2)
         @Comment("How rare are Nether Brick Outposts in non-warped Nether biomes."
                 + "\n1 for spawning in most chunks and 1001 for none.")