+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSStrongholdsConfig.java
@@ -26,6 +26,12 @@ public class RSStrongholdsConfig implements ConfigData
 	@Comment("Controls whether loot chests spawn or not in the Stronghold.")
 	public boolean lootChestsSH = true;
 
+	@ConfigEntry.Gui.Tooltip(count = 3)
+	@Comment("Add the ID/resource location of the biome you don't want"
+			+"\nRS's strongholds to spawn in. Separate each ID with a comma ,"
+			+"\n"
+			+"\nExample: \"minecraft:ice_spikes,awesome_mod:awesome_biome\"")
+	public String blacklistedStrongholdBiomes = "";
 
 	@ConfigEntry.Gui.CollapsibleObject
 	public RSStrongholdsConfig.Stonebrick stonebrick = new RSStrongholdsConfig.Stonebrick();