+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSTemplesConfig.java
@@ -17,6 +17,13 @@ public class RSTemplesConfig implements ConfigData {
 
     public static class Temples {
 
+        @ConfigEntry.Gui.Tooltip(count = 3)
+        @Comment("Add the ID/resource location of the biome you don't want"
+                +"\nRS's temples to spawn in. Separate each ID with a comma ,"
+                +"\n"
+                +"\nExample: \"minecraft:ice_spikes,awesome_mod:awesome_biome\"")
+        public String blacklistedTempleBiomes = "";
+
         @ConfigEntry.Gui.Tooltip(count = 2)
         @Comment("How rare are Nether Temples in Nether Wastelands."
                 + "\n1 for spawning in most chunks and 1001 for none.")
@@ -74,6 +81,13 @@ public class RSTemplesConfig implements ConfigData {
 
 
     public static class Pyramids {
+
+        @ConfigEntry.Gui.Tooltip(count = 3)
+        @Comment("Add the ID/resource location of the biome you don't want"
+                +"\nRS's pyramids to spawn in. Separate each ID with a comma ,"
+                +"\n"
+                +"\nExample: \"minecraft:ice_spikes,awesome_mod:awesome_biome\"")
+        public String blacklistedPyramidBiomes = "";
 
         @ConfigEntry.Gui.Tooltip(count = 2)
         @Comment("How rare are Nether Pyramids in Nether."